moving scrolling container and image scrollbar
