+++ scripts/game_structure/ui_elements.py
@@ -2,7 +2,6 @@ import math
 from functools import lru_cache
 from math import ceil
 from typing import (
-    Tuple,
     Optional,
     List,
     Union,
@@ -13,8 +12,8 @@ import i18n
 import pygame
 import pygame_gui
 from pygame_gui.core import UIContainer, IContainerLikeInterface, UIElement, ObjectID
-from pygame_gui.core.gui_type_hints import RectLike, Coordinate
-from pygame_gui.core.interfaces import IUIManagerInterface, IUIElementInterface
+from pygame_gui.core.gui_type_hints import RectLike
+from pygame_gui.core.interfaces import IUIManagerInterface
 from pygame_gui.elements import UIAutoResizingContainer
 
 from scripts.cat_relations.enums import RelType, RelTier
@@ -24,6 +23,9 @@ from scripts.game_structure import game
 from scripts.game_structure.screen_settings import MANAGER
 from scripts.game_structure.game.settings import game_setting_get
 from scripts.ui.elements.image_button import UIImageButton
+from scripts.ui.elements.modified_scrolling_container import (
+    UIModifiedScrollingContainer,
+)
 from scripts.ui.elements.sprite_button import UISpriteButton
 from scripts.ui.elements.surface_image_button import UISurfaceImageButton
 from scripts.ui.elements.text_box_tweaked import UITextBoxTweaked
@@ -32,313 +34,6 @@ from scripts.ui.generate_button import get_button_dict, ButtonStyles
 from scripts.ui.icon import Icon
 from scripts.events_module.text_adjust import shorten_text_to_fit
 from scripts.ui.scale import ui_scale, ui_scale_dimensions, ui_scale_value
-
-
-class UIModifiedScrollingContainer(
-    pygame_gui.elements.UIScrollingContainer, IContainerLikeInterface
-):
-    def __init__(
-        self,
-        relative_rect: pygame.Rect,
-        manager=None,
-        starting_height: int = 1,
-        container=None,
-        object_id=None,
-        visible: int = 1,
-        allow_scroll_x: bool = False,
-        allow_scroll_y: bool = False,
-        should_grow_automatically=True,
-        anchors=None,
-    ):
-        self.scroll_bar_starting_height = 1
-
-        super().__init__(
-            relative_rect=relative_rect,
-            manager=manager,
-            starting_height=starting_height,
-            container=container,
-            object_id=object_id,
-            visible=visible,
-            allow_scroll_x=allow_scroll_x,
-            allow_scroll_y=allow_scroll_y,
-            should_grow_automatically=should_grow_automatically,
-            anchors=anchors,
-        )
-        self.scroll_bar_starting_height = self.get_top_layer()
-        if self.allow_scroll_y:
-            self.vert_scroll_bar.kill()
-            self.vert_scroll_bar = None
-
-            self.scroll_bar_width = self._get_scroll_bar_width()
-            scroll_bar_rect = pygame.Rect(
-                -self.scroll_bar_width,
-                0,
-                self.scroll_bar_width,
-                self.relative_rect.height,
-            )
-
-            self.vert_scroll_bar = UIImageVerticalScrollBar(
-                relative_rect=scroll_bar_rect,
-                visible_percentage=1.0,
-                manager=self.ui_manager,
-                container=self._root_container,
-                parent_element=self,
-                starting_height=self.scroll_bar_starting_height,
-                anchors={
-                    "left": "right",
-                    "right": "right",
-                    "top": "top",
-                    "bottom": "bottom",
-                },
-                visible=False,
-            )
-            self.join_focus_sets(self.vert_scroll_bar)
-
-            self.vert_scroll_bar.set_container_this_will_scroll(
-                self.scrollable_container
-            )
-
-        if self.allow_scroll_x:
-            self.horiz_scroll_bar.kill()
-            self.horiz_scroll_bar = None
-
-            self.scroll_bar_height = self._get_scroll_bar_height()
-
-            scroll_bar_rect = pygame.Rect(
-                0,
-                -self.scroll_bar_height,
-                self.relative_rect.width,
-                self.scroll_bar_height,
-            )
-            self.horiz_scroll_bar = UIModifiedHorizScrollBar(
-                relative_rect=scroll_bar_rect,
-                visible_percentage=1.0,
-                manager=self.ui_manager,
-                container=self._root_container,
-                parent_element=self,
-                anchors={
-                    "left": "left",
-                    "right": "right",
-                    "top": "bottom",
-                    "bottom": "bottom",
-                },
-                visible=True,
-                starting_height=self.scroll_bar_starting_height,
-            )
-            self.horiz_scroll_bar.set_dimensions((self.relative_rect.width, 0))
-            self.horiz_scroll_bar.set_relative_position((0, 0))
-            self.horiz_scroll_bar.set_container_this_will_scroll(
-                self.scrollable_container
-            )
-            self.join_focus_sets(self.horiz_scroll_bar)
-
-    def set_view_container_dimensions(self, dimensions: Coordinate):
-        self._view_container.set_dimensions(dimensions)
-
-    def set_dimensions(self, dimensions, clamp_to_container: bool = False):
-        super().set_dimensions(dimensions, clamp_to_container)
-
-    def on_contained_elements_changed(self, target: IUIElementInterface) -> None:
-        """
-        Update the positioning of the contained elements of this container. To be called when one of the contained
-        elements may have moved, been resized or changed its anchors.
-
-        :param target: the UI element that has been benn moved resized or changed its anchors.
-        """
-        self.scrollable_container.on_contained_elements_changed(target)
-
-    def _sort_out_element_container_scroll_bars(self):
-        """
-        This creates, re-sizes or removes the scrollbars after resizing, but not after the scroll
-        bar has been moved. Instead, it tries to keep the scrollbars in the same approximate position
-        they were in before resizing
-        """
-        self.scroll_bar_width = self._get_scroll_bar_width()
-        super()._sort_out_element_container_scroll_bars()
-
-        if self.vert_scroll_bar:
-            self.vert_scroll_bar.change_layer(self.scroll_bar_starting_height)
-            self.vert_scroll_bar.show()
-
-        if self.horiz_scroll_bar:
-            self.horiz_scroll_bar.change_layer(self.scroll_bar_starting_height)
-            self.horiz_scroll_bar.show()
-
-    def _check_scroll_bars(self) -> Tuple[bool, bool]:
-        """
-        Check if we need a horizontal or vertical scrollbar.
-        """
-        self.scroll_bar_width = 0
-        self.scroll_bar_height = 0
-        need_horiz_scroll_bar = False
-        need_vert_scroll_bar = False
-
-        if (
-            self.scrolling_height > self._view_container.rect.height
-            or self.scrollable_container.relative_rect.top != 0
-        ) and self.allow_scroll_y:
-            need_vert_scroll_bar = True
-            self.scroll_bar_width = self._get_scroll_bar_width()
-
-        # Need to subtract scrollbar width here to account for when the above statement evaluated to True
-        if (
-            self.scrolling_width
-            > self._view_container.rect.width - self.scroll_bar_width
-            or self.scrollable_container.relative_rect.left != 0
-        ) and self.allow_scroll_x:
-            need_horiz_scroll_bar = True
-            self.scroll_bar_height = self._get_scroll_bar_height()
-
-            # Needs a second check for the case where we didn't need the vertical scroll bar until after creating a
-            # horizontal scroll bar
-            if (
-                self.scrolling_height
-                > self._view_container.rect.height - self.scroll_bar_height
-                or self.scrollable_container.relative_rect.top != 0
-            ) and self.allow_scroll_y:
-                need_vert_scroll_bar = True
-                self.scroll_bar_width = self._get_scroll_bar_width()
-
-        self._calculate_scrolling_dimensions()
-        return need_horiz_scroll_bar, need_vert_scroll_bar
-
-    def _get_scroll_bar_width(self) -> int:
-        return ui_scale_value(20) + 4
-
-    def _get_scroll_bar_height(self) -> int:
-        return ui_scale_value(18) + 2
-
-    def are_contents_hovered(self) -> bool:
-        """
-        Are any of the elements in the container hovered? Used for handling mousewheel events.
-
-        :return: True if one of the elements is hovered, False otherwise.
-        """
-        for element in self:
-            if any(sub_element.hovered for sub_element in element.get_focus_set()):
-                return True
-            elif (
-                isinstance(element, IContainerLikeInterface)
-                and element.are_contents_hovered()
-            ):
-                return True
-        return False
-
-
-class UIImageVerticalScrollBar(pygame_gui.elements.UIVerticalScrollBar):
-    def __init__(
-        self,
-        relative_rect: pygame.Rect,
-        visible_percentage: float,
-        manager=None,
-        container=None,
-        parent_element=None,
-        object_id=None,
-        anchors=None,
-        visible: int = 1,
-        starting_height: int = 1,
-    ):
-        super().__init__(
-            relative_rect=relative_rect,
-            visible_percentage=visible_percentage,
-            manager=manager,
-            container=container,
-            parent_element=parent_element,
-            object_id=object_id,
-            anchors=anchors,
-            visible=visible,
-        )
-
-        self.scroll_wheel_speed = 100
-        self.sliding_button.change_layer(starting_height)
-        self.button_height = 16
-        self.arrow_button_height = self.button_height
-        self.top_button.kill()
-        self.top_button = UIImageButton(
-            ui_scale(pygame.Rect((0, 0), (16, 16))),
-            text="",
-            manager=self.ui_manager,
-            container=self.button_container,
-            starting_height=starting_height,
-            parent_element=self,
-            object_id="#vertical_slider_up_arrow_button",
-            anchors={
-                "centerx": "centerx",
-            },
-        )
-
-        self.bottom_button.kill()
-        bottom_button_rect = ui_scale(pygame.Rect((0, 0), (16, 16)))
-        bottom_button_rect.bottomleft = (0, 0)
-        self.bottom_button = UIImageButton(
-            bottom_button_rect,
-            text="",
-            manager=self.ui_manager,
-            container=self.button_container,
-            starting_height=starting_height,
-            parent_element=self,
-            object_id="#vertical_slider_down_arrow_button",
-            anchors={
-                "bottom": "bottom",
-                "centerx": "centerx",
-            },
-        )
-        del bottom_button_rect
-
-        self.sliding_button.kill()
-        scroll_bar_height = max(
-            5, int(self.scrollable_height * self.visible_percentage)
-        )
-        self.sliding_button = pygame_gui.elements.UIButton(
-            pygame.Rect(
-                (
-                    int(self.sliding_rect_position[0]),
-                    int(self.sliding_rect_position[1]),
-                ),
-                (self.background_rect.width, scroll_bar_height),
-            ),
-            "",
-            self.ui_manager,
-            container=self.button_container,
-            starting_height=starting_height,
-            parent_element=self,
-            object_id="#sliding_button",
-            anchors={"left": "left", "right": "right", "top": "top", "bottom": "top"},
-        )
-
-        self.join_focus_sets(self.sliding_button)
-        self.sliding_button.set_hold_range((100, self.background_rect.height))
-
-    def set_visible_percentage(self, percentage: float):
-        super().set_visible_percentage(percentage)
-        self.scroll_wheel_speed = (1 / self.visible_percentage) * ui_scale_value(15)
-
-    def _check_should_handle_mousewheel_event(self) -> bool:
-        def recursive_check_if_ignore(element):
-            """
-            If this is TRUE, we should ignore the scroll. This just helps with shortcutting
-            :param element: The UIElement to check
-            :return: True to ignore, False if we should care
-            """
-            if (
-                isinstance(element, (UIScrollingDropDown, UIScrollingButtonList))
-                and element.are_contents_hovered()
-            ):
-                return True
-            elif isinstance(element, IContainerLikeInterface):
-                for sub_element in element:
-                    if recursive_check_if_ignore(sub_element):
-                        return True
-            return False
-
-        # inverting the outcome of that
-        if any(recursive_check_if_ignore(ele) for ele in self._container_to_scroll):
-            return False
-        else:
-            return (
-                self._container_to_scroll
-                and self._container_to_scroll.are_contents_hovered()
-            ) or self._check_is_focus_set_hovered()
 
 
 class UIModifiedHorizScrollBar(pygame_gui.elements.UIHorizontalScrollBar):
