+++ scripts/screens/ClanSettingsScreen.py
@@ -12,9 +12,9 @@ import ujson
 from scripts.cat.cats import Cat
 from scripts.game_structure.game.settings import game_setting_get
 from scripts.game_structure.ui_elements import (
-    UIModifiedScrollingContainer,
     UICheckbox,
 )
+from ..ui.elements.modified_scrolling_container import UIModifiedScrollingContainer
 from ..ui.elements.image_button import UIImageButton
 from ..ui.elements.surface_image_button import UISurfaceImageButton
 from scripts.housekeeping.datadir import open_data_dir
