+++ scripts/screens/EventEditScreen.py
@@ -21,13 +21,15 @@ from scripts.game_structure import game
 from scripts.cat.pronouns import get_default_pronouns
 from scripts.game_structure.screen_settings import MANAGER
 from scripts.game_structure.ui_elements import (
-    UIModifiedScrollingContainer,
     UICheckbox,
     UIModifiedImage,
     UIScrollingButtonList,
     UIDropDown,
     UICollapsibleContainer,
     UIScrollingDropDown,
+)
+from scripts.ui.elements.modified_scrolling_container import (
+    UIModifiedScrollingContainer,
 )
 from scripts.ui.elements.text_box_tweaked import UITextBoxTweaked
 from scripts.ui.elements.surface_image_button import UISurfaceImageButton
