+++ scripts/ui/windows/symbol_filter.py
@@ -8,7 +8,7 @@ from scripts.game_structure.game.switches import (
     switch_remove_list_value,
 )
 from scripts.game_structure.screen_settings import MANAGER
-from scripts.game_structure.ui_elements import (
+from scripts.ui.elements.modified_scrolling_container import (
     UIModifiedScrollingContainer,
 )
 from scripts.ui.elements.image_button import UIImageButton
